requirements.txt
qrcode==7.4
pillow==11.2.1
numpy==2.2.5
streamlit==1.44.1
matplotlib==3.10.3